+++ Hello/main.py
@@ -32,15 +32,15 @@ import atlastk as Atlas
 def readAsset(path):
 	return Atlas.readAsset(path, "Hello")
 
-def acConnect(self, dom, id):
+def acConnect(self, dom):
 	dom.setLayout("", readAsset( "Main.html") )
 	dom.focus( "input")
 
-def acSubmit(self, dom, id):
+def acSubmit(self, dom):
 	dom.alert("Hello, " + dom.getContent("input") + "!")
 	dom.focus( "input")
 
-def acClear(self, dom, id):
+def acClear(self, dom):
 	if ( dom.confirm("Are you sure?" ) ):
 		dom.setContent("input", "" )
 	dom.focus( "input")