+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 - Common to **all** bindings,
 - does **not** concern the native code,
 
+## 2019-07-01
+
+- *Python*:
+  - the callbacks can now have none to four (user object, *DOM* object, id, action) parameters,
+  - switching to 0.9.8
+
 ## 2019-06-01
 
 - *Python*:
@@ -20,8 +26,8 @@
 ## 2019-05-09
 
 - *Node.js*:
-	- changing some stuff related to *Repl.it*,
-	- switching to 0.9.2,
+  - changing some stuff related to *Repl.it*,
+  - switching to 0.9.2,
 
 ## 2019-05-07